Share some maplibre util code, hook up zorder, and fix streetview in browse
